blöder Benny

+++ tfp-collab-userspace-infrastructure-database/Repositories/GalleryRepository.cs
@@ -1,7 +1,6 @@
 using System.Data;
 using Dapper;
 using FluentResults;
-using Microsoft.Extensions.Logging;
 using tfp_collab_userspace_domain.DomainObjects;
 using tfp_collab_userspace_domain.Service;
 using tfp_collab_userspace_storage_database.Mapping;
@@ -21,6 +20,8 @@ public class GalleryRepository
 
     public async Task<Result<GalleryDo>> GetAsync(OwnerId ownerId, GalleryId galleryId)
     {
+        if (_transaction.Connection is null) throw new InvalidOperationException("Connection is null.");
+        
         try
         {
             var gallery = 
@@ -28,8 +29,7 @@ public class GalleryRepository
                     "SELECT * FROM Gallery WHERE OwnerId = @OwnerId and Id = @Id", 
                     new { Id = galleryId.Value, OwnerId =  ownerId.Value });
             
-            if (gallery is null) return Result.Fail("Gallery has not been selected.");
-            return Result.Ok(gallery.ToDo());
+            return gallery is null ? Result.Fail("Gallery has not been selected.") : Result.Ok(gallery.ToDo());
         }
         catch (Exception ex)
         {
@@ -39,14 +39,15 @@ public class GalleryRepository
     
     public async Task<Result<IEnumerable<GalleryDo>>> GetAsync(OwnerId ownerId)
     {
+        if (_transaction.Connection is null) throw new InvalidOperationException("Connection is null.");
+        
         try
         {
             var galleries = 
                 await _transaction.Connection
                     .QueryAsync<Gallery>("SELECT * FROM Gallery WHERE OwnerId = @ownerId and");
             
-            if (galleries is null) return Result.Fail("Galleries cannot be selected.");
-            return Result.Ok(galleries.Select(g => g.ToDo()));
+            return !galleries.Any() ? Result.Fail("Galleries cannot be selected.") : Result.Ok(galleries.Select(g => g.ToDo()));
         }
         catch (Exception ex)
         {
@@ -56,19 +57,17 @@ public class GalleryRepository
 
     public async Task<Result<GalleryDo>> CreateAsync(GalleryDo galleryDo)
     {
+        if (_transaction.Connection is null) throw new InvalidOperationException("Connection is null.");
         try
         {
-            Gallery? gallery = galleryDo.ToModel();
-            if (gallery is null)
-            {
-                
-                return Result.Fail("Gallery has not been created.");
-            }
+            var gallery = galleryDo.ToModel();
+            
             gallery.Id = GalleryId.New();
             gallery.AddedOn = DateTime.UtcNow;
-            var rowsInserted = await _transaction.Connection.ExecuteAsync(
-                "INSERT INTO Gallery (Id, OwnerId, Name, AddedOn) VALUES (@Id, @OwnerId, @Name, @AddedOn)", 
-                gallery);
+            var rowsInserted = 
+                await _transaction.Connection.ExecuteAsync(
+                    "INSERT INTO Gallery (Id, OwnerId, Name, AddedOn) VALUES (@Id, @OwnerId, @Name, @AddedOn)", 
+                    gallery);
             
             if (rowsInserted <= 0) return Result.Fail("Gallery has not been created.");
             galleryDo = gallery.ToDo();
@@ -82,14 +81,16 @@ public class GalleryRepository
 
     public async Task<Result> DeleteAsync(GalleryId id)
     {
+        if (_transaction.Connection is null) throw new InvalidOperationException("Connection is null.");
+        
         try
         {
-            var rowsDeleted = await _transaction.Connection.ExecuteAsync(
-                "DELETE FROM Gallery WHERE Id = @Id", 
-                new { Id = id.Value });
+            var rowsDeleted = 
+                await _transaction.Connection.ExecuteAsync(
+                    "DELETE FROM Gallery WHERE Id = @Id", 
+                    new { Id = id.Value });
             
-            if (rowsDeleted <= 0) return Result.Fail("Gallery has not been deleted.");
-            return Result.Ok();
+            return rowsDeleted <= 0 ? Result.Fail("Gallery has not been deleted.") : Result.Ok();
         }
         catch (Exception ex)
         {
